add tripEvents controller
