+++ bigTrip/src/controllers/types.ts
@@ -1,0 +1,3 @@
+export interface Controller {
+  _container: Element | null;
+}
